+++ changelog.md
@@ -4,6 +4,7 @@ Beta
 0.1.10 -> 0.1.11
 
 
+- Grabbing system to allow Titans to pick up smaller entities
 - Fixed titan combat
 - Fixed armored titan charging wackiness
 - Perfect Parrying attacks to deal counter damage
@@ -12,6 +13,7 @@ Beta
 - Added Titan footstep sounds
 - Tweaked sprint charging so it only uses stamina if you've shed enough armor
 - Added a few sound effects for titan combat and odm blade swinging
+- General bug fixes and smaller tweaks
 
 TODO:
 - Investigate why blade handles stop swinging randomly